+++ .gitignore
@@ -1,5 +1,6 @@
 bin/
 *.iso
+*.img
 
 *.icloud
 .DS_Store
