+++ src/system.h
@@ -14,6 +14,8 @@
         _assert_1(__VA_ARGS__),\
         _assert_0(__VA_ARGS__))
 
+const char *get_notification();
+void notify(const char *err);
 void panic(const char *err);
 u32 rand();
 void seed(u32 s);
